refactor: `apis/` -> `api/`
