+++ src/background/messageListeners/api/favorite.ts
