+++ src/utils/api.ts
@@ -1,4 +1,4 @@
-import type { API_COLLECTION } from '~/background/messageListeners/apis'
+import type { API_COLLECTION } from '~/background/messageListeners/api'
 
 type CamelCase<S extends string> = S extends `${infer P1}_${infer P2}${infer P3}`
   ? `${Lowercase<P1>}${Uppercase<P2>}${CamelCase<P3>}`
